# perf: Add benches for parser and get_path