--- a/Cargo.toml
+++ b/Cargo.toml
@@ -36,4 +36,18 @@ serde_json = { version = "1.0.95", default-features = false, features = [
 ] }
 
 [dev-dependencies]
-goldenfile = "1.4.5"
+goldenfile = "1.5.2"
+serde_json = "1.0.105"
+json-deserializer = "0.4.4"
+simd-json = {version = "0.10.6", features = ["allow-non-simd"]}
+mockalloc = "0.1.2"
+criterion = "0.5.1"
+
+[[bench]]
+name = "parser"
+harness = false
+
+[[bench]]
+name = "get_path"
+harness = false
+