parser benches

| canada.json       |                                 |
| ----------------- | ------------------------------- |
| jsonb             | [4.6569 ms 4.6622 ms 4.6680 ms] |
| serde_json        | [8.7727 ms 8.7911 ms 8.8115 ms] |
| json_deserializer | [4.5349 ms 4.5409 ms 4.5471 ms] |
| simd_json         | [5.1894 ms 5.1993 ms 5.2100 ms] |

| citm_catalog.json |                                 |
| ----------------- | ------------------------------- |
| jsonb             | [2.5120 ms 2.5162 ms 2.5206 ms] |
| serde_json        | [4.0064 ms 4.0127 ms 4.0197 ms] |
| json_deserializer | [2.9674 ms 2.9717 ms 2.9761 ms] |
| simd_json         | [2.1907 ms 2.1936 ms 2.1968 ms] |

| twitter.json      |                                 |
| ----------------- | ------------------------------- |
| jsonb             | [1.9177 ms 1.9210 ms 1.9245 ms] |
| serde_json        | [2.1374 ms 2.1403 ms 2.1435 ms] |
| json_deserializer | [2.1390 ms 2.1418 ms 2.1448 ms] |
| simd_json         | [708.68 µs 710.22 µs 711.88 µs] |


get path benches
|                                      | jsonb                           | serde_json                      |
| ------------------------------------ | ------------------------------- | ------------------------------- |
| canada->type                         | [91.082 ns 91.353 ns 91.623 ns] | [10.425 ms 10.447 ms 10.472 ms] |
| citm_catalog->areaNames->205705994   | [176.80 ns 177.42 ns 178.07 ns] | [4.0597 ms 4.0674 ms 4.0759 ms] |
| citm_catalog->topicNames->324846100  | [159.70 ns 160.09 ns 160.47 ns] | [4.1255 ms 4.1321 ms 4.1388 ms] |
| twitter->search_metadata->max_id_str | [141.68 ns 142.00 ns 142.28 ns] | [2.1449 ms 2.1493 ms 2.1541 ms] |

Closes #28